Decompose AppRoot into focused files
Split the monolithic AppRoot.kt (312 lines, 10+ functions) into three
focused files:

- AppRoot.kt (97 lines): Entry point + theme/reset helpers
- AppContent.kt: Content routing, loading screen, mode sync
- FeaturesState.kt: Feature lifecycle management and state

Removes @file:Suppress("TooManyFunctions") and simplifies
createResetAppCallback into a direct resetApp function.
